IRAF PCIX Version 2.16
 * Enhanced @-file templates - Expansion of MEF files and tables
 * New VO-CL Command Language - New VO builtin functions
 * URL Support in all tasks - Access to remote data
 * VOTable Support in all tasks - Transparent XML support
 * Application interoperability - Desktop messaging
 * New VO External Package - New VO Toolbox
 * Removal of all license restrictions - IRAF is now free

+++ unix/os/zmaloc.c
@@ -24,7 +24,7 @@ ZMALOC (
 	register char *bufptr;
 	int      stat;
 
-	bufptr = malloc ((int)*nbytes);
+	bufptr = malloc ((size_t)*nbytes);
 	if (bufptr != NULL) {
 	    *buf = ADDR_TO_LOC(bufptr);
             if (*buf > 0)
